move functions to utils
