+++ utils.py
@@ -49,3 +49,41 @@ def get_tracks_list(sp_rqrmt, url_playlist="https://open.spotify.com/collection/
         all_tracks = all_tracks[:max_tracks]
 
     return all_tracks
+
+
+#Function to extract features
+def extract_features(img_path, model):
+    img=image.load_img(img_path, target_size=(224,224)) #TODO: check size of album covers
+    # img=img.convert("RGB")
+    img_array = image.img_to_array(img)
+    img_array = np.expand_dims(img_array, axis=0)
+    img_array = preprocess_input(img_array)
+    features = model.predict(img_array)
+    return features.flatten()
+
+
+def correlation_similarity(vector1, vector2):
+    """
+    Calculate correlation similarity between two vectors.
+    
+    Args:
+    Return:
+    
+    """
+    # Ensure the vectors are NumPy arrays
+    vector1 = np.array(vector1)
+    vector2 = np.array(vector2)
+    
+    correlation_scores = []
+    
+    # Calculate correlation coefficient
+    if len(vector2) > 1:
+        for vector in vector2:
+            correlation_scores.append(np.corrcoef(vector1, vector)[0, 1])
+    else:
+        correlation_scores.append(np.corrcoef(vector1, vector2)[0, 1])
+    return correlation_scores
+
+def normalize_feature(features):
+    norms = np.linalg.norm(features, axis=1, keepdims=True)
+    return features/norms
